Implementor:   new: exceptions, constructors processing, Melnikov test   fixed: arguments processing Bag:   new: optimized clearAll   moved: Melnikov tests

+++ src/ru/ifmo/ctddev/rybak/reflect/test/Int1.java
@@ -1,6 +1,5 @@
 package ru.ifmo.ctddev.rybak.reflect.test;
 public interface Int1 {
 	public Object i();
-
-	public int OverRided();
+	public float OverRided();
 }
