+++ src/ru/ifmo/ctddev/rybak/reflect/test/Int2.java
@@ -1,6 +1,7 @@
 package ru.ifmo.ctddev.rybak.reflect.test;
+
 public interface Int2 {
 	public Integer i();
 
-	public int OverRided();
+	public float OverRided();
 }
